[WebUI] Fixed a crash when editing Subscribe
After the UE performs Registration/Attach, SQN field is created.

If we edit subscriber information when SQN value is present, WebUI crash occurs.

It is because the way to handle Long Type(SQN:Long) is different
when the mongoose version is 6 or higher.

To avoid this crash, we use the mongoose version down to 5.x first.